backend/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
docker-py==1.10.6